+++ crates/hyperswitch_connectors/src/connectors/celero/transformers.rs
@@ -191,6 +191,7 @@ impl TryFrom<&PaymentMethodData> for CeleroPaymentMethod {
             }
             PaymentMethodData::CardDetailsForNetworkTransactionId(_)
             | PaymentMethodData::NetworkTokenDetailsForNetworkTransactionId(_)
+            | PaymentMethodData::CardWithLimitedDetails(_)
             | PaymentMethodData::CardRedirect(_)
             | PaymentMethodData::Wallet(_)
             | PaymentMethodData::PayLater(_)
@@ -323,7 +324,8 @@ fn determine_cit_mit_fields(
         }
         // For other mandate types that might not be supported
         Some(api_models::payments::MandateReferenceId::NetworkMandateId(_))
-        | Some(api_models::payments::MandateReferenceId::NetworkTokenWithNTI(_)) => {
+        | Some(api_models::payments::MandateReferenceId::NetworkTokenWithNTI(_))
+        | Some(api_models::payments::MandateReferenceId::CardWithLimitedData) => {
             // These might need different handling or return an error
             Err(errors::ConnectorError::NotImplemented(
                 get_unimplemented_payment_method_error_message("Celero"),